requirement.txt
flask
fastapi
uvicorn
httpx
pycryptodome
protobuf